requirements.txt
streamlit
langchain
googlegenerativeai
